Total change.
For this C first program first thing the translation wasn't correct , and this mistakes makes the Hello World tutorial/course not even understandable. So I decide to keep on the same structure from A to Z, but I did change everything else; explanation, styling, and also I did added Pictures that shows example of what I explain above so it make it easier for anyone to understand.
And the plus extra thing is adding a second fonction, the write() fonction that not very popular but one of the important fonctions that a programmer should be knowing.